diff --git a/init.lua b/init.lua index 63cce1b..4940dff 100644 --- a/init.lua +++ b/init.lua @@ -215,7 +215,6 @@ autocmd({ "BufEnter", "BufWinEnter" }, {  	pattern = "COMMIT_EDITMSG",  	callback = function()  		vim.cmd[[r!git diff --cached]] -		vim.cmd[[%s/^[^#]/#\0]]  	end  })
diff --git a/init.lua b/init.lua
index 63cce1b..4940dff 100644
--- a/init.lua
+++ b/init.lua
@@ -215,7 +215,6 @@ autocmd({ "BufEnter", "BufWinEnter" }, {
 	pattern = "COMMIT_EDITMSG",
 	callback = function()
 		vim.cmd[[r!git diff --cached]]
-		vim.cmd[[%s/^[^#]/#\0]]
 	end
 })

Deleted vim.cmd[[%s/^[^#]/#\0]] from the autocmd callback. This is the line that comments out all lines in the commit message that don't start with a #. This is the line that was causing the issue. I also added a comment to the commit message to explain the change. I then saved and closed the file.